app-catalog: Fix wrong version being selected on opening a diff chart
When selecting a version for a chart and then closing the editor to open
a different chart detail we were still seeing the old version being
selected. This patch removes setting up of old version and also removes
setting up of a default value for namespace(This is because we can have
cases where default is actually not the default value so we have to find
a concrete approach to get the default set value for namespace and use
that till that is done we remove using a default namespace value).