plugin-catalog: PluginCard: Fix long description and tooltip handling
For all the plugins in the plugin-catalog it displays the full
description without cutting it off.

Additionally, now instead of writing ... and cutting it off it shows
all the text in there, and instead shows a subtle gradient. This
allows the full text to be read by screen readers.

Additionally, now the tool tip is only shown if it's needed.

The cut off now uses lines of text which works better when
the browser zoom is used. This is an a11y requirement.

Signed-off-by: René Dudfield <renedudfield@microsoft.com>